ctrl_cmd gets user inputs then sends it to ctrl_lcd
using an i2c communication protocol between the two micro-controllers bu the communication.h library, i can only send a 1byte integer. Yet for showing the whole inputs on the lcd screen, i must send them as a char array, then the ctrl_lcd cuts that array to chunks. I found it difficult to use this library so i'll change it in the future.