Added booting animation and ditched the idea of lcd controller sending data via nRF
The welcome functions runs at starting up and to draw a bitmap animation stored in PROGMEM to reduce memory consumption (although i'm beyond 90% memory and 90% storage). After the animation it greets the user with "READY TO TAKE OFF" message.

Removed code responsible for sending the user Inputs to the drone flight controller.

The lcd controller responsible only for the screen now.